🎨: デザインシステム連携のため企業アカウント選択パターンの記述構成を見直す
